+++ src/Core/Metrics/StatsDMetricsClient.cs
@@ -31,7 +31,7 @@ namespace Foundatio.Metrics {
             return TaskHelper.Completed();
         }
 
-        public Task TimerAsync(string statName, long milliseconds) {
+        public Task TimerAsync(string statName, int milliseconds) {
             Send(BuildMetric("ms", statName, milliseconds.ToString(CultureInfo.InvariantCulture)));
             return TaskHelper.Completed();
         }
