+++ src/MetricsNet/MetricsNETClient.cs
@@ -16,7 +16,7 @@ namespace Foundatio.Metrics {
             return TaskHelper.Completed();
         }
 
-        public Task TimerAsync(string statName, long milliseconds) {
+        public Task TimerAsync(string statName, int milliseconds) {
             Metric.Timer(statName, Unit.Calls, SamplingType.SlidingWindow, TimeUnit.Milliseconds).Record(milliseconds, TimeUnit.Milliseconds);
             return TaskHelper.Completed();
         }
